Interal cmd. Config edition!
Outwards facing configuration with a lot of customisation for the /the_vault internal command.

(GIFT_SUB/DONOR/BIT)_BRACKETS are in order of "lowest" to "highest", -1 is "disabled".

(DONOR/BIT)_MEGAHEAD is the value to reach for the generated trader to be a "MegaHead"

(GIFT_SUB/SUB)_GIVE_BITS gives the Bit-equivalent in dollars. (Aka a Prime/$5 sub is 500 bits.) - This also promotes to a MEGAHEAD if the amount is 2500 for a SUB only

SUB_BRACKET_AS_RARITY upgrades the rarity of the trader based on the tier of the sub (prime/normal as a "common", $10 as a "rare", $25 as a "OMEGA"

NOTE:
The gift_bomb.json is basically un-opulated of items so any donations won't really generate items.
The trader_corE_omega.json is also full of trades with "0" available and will need adjusting.